requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via discord-py
aiosignal==1.4.0
    # via aiohttp
attrs==25.3.0
    # via aiohttp
audioop-lts==0.2.2
    # via discord-py
cffi==2.0.0
    # via pynacl
colorama==0.4.6
    # via loguru
discord-py==2.6.3
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.10
    # via yarl
loguru==0.7.3
    # via -r requirements.in
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pycparser==2.23
    # via cffi
pynacl==1.5.0
    # via discord-py
python-dotenv==1.1.1
    # via -r requirements.in
ruff==0.13.1
    # via -r requirements.in
win32-setctime==1.2.0
    # via loguru
yarl==1.20.1
    # via aiohttp
